# swag-docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.CreateShardRequest:
    properties:
      nodes:
        items:
          type: string
        type: array
      password:
        type: string
    required:
    - nodes
    type: object
  api.FailoverRequest:
    properties:
      preferred_node_id:
        type: string
    type: object
  api.FailoverResponse:
    properties:
      new_master_id:
        type: string
    type: object
  api.GetShardResponse:
    properties:
      shard:
        $ref: '#/definitions/store.Shard'
    type: object
...
  api.ListShardResponse:
    properties:
      shards:
        items:
          $ref: '#/definitions/store.Shard'
        type: array
    type: object
...
  helper.Error:
    properties:
      message:
        type: string
    type: object
  helper.Response:
    description: General Response Structure
    properties:
      data: {}
      error:
        $ref: '#/definitions/helper.Error'
    type: object
...
  store.Shard:
    description: Node role
    properties:
      migrating_slot:
        type: integer
      nodes:
        items: {}
        type: array
      slot_ranges:
        items:
          $ref: '#/definitions/store.SlotRange'
        type: array
      target_shard_index:
        type: integer
    type: object
  store.SlotRange:
    properties:
      start:
        type: integer
      stop:
        type: integer
    type: object
info:
  contact: {}
  title: Kvrocks Controller API
  version: "1.0"
paths:
...
  /namespaces/{namespace}/clusters/{cluster}/shards:
    get:
      description: List all shards
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Cluster
        in: path
        name: cluster
        required: true
        type: string
      responses:
        "200":
          description: 分片列表
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.ListShardResponse'
              type: object
      summary: List shards
      tags:
      - shard
    post:
      consumes:
      - application/json
      description: Create a shard by providing nodes
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Cluster
        in: path
        name: cluster
        required: true
        type: string
      - description: Create Shard Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.CreateShardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/store.Shard'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Error'
      summary: Create a shard
      tags:
      - shard
  /namespaces/{namespace}/clusters/{cluster}/shards/{shard}:
    delete:
      consumes:
      - application/json
      description: Remove a shard by providing shard index
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Cluster
        in: path
        name: cluster
        required: true
        type: string
      - description: Shard Index
        in: path
        name: shard
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Error'
      summary: Remove a shard
      tags:
      - shard
    get:
      description: Get a shard by providing shard index
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Cluster
        in: path
        name: cluster
        required: true
        type: string
      - description: Shard Index
        in: path
        name: shard
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.GetShardResponse'
              type: object
      summary: Get a shard
      tags:
      - shard
  /namespaces/{namespace}/clusters/{cluster}/shards/{shard}/failover:
    post:
      consumes:
      - application/json
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Cluster
        in: path
        name: cluster
        required: true
        type: string
      - description: Shard Index
        in: path
        name: shard
        required: true
        type: integer
      - description: Failover Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.FailoverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.FailoverResponse'
              type: object
      summary: Execute failover on a target shard
      tags:
      - shard
...
swagger: "2.0"